refactor files around into appropriate search directory
