+++ datahub-web-react/src/app/entity/dashboard/DashboardEntity.tsx
@@ -24,7 +24,7 @@ import { EntityMenuItems } from '../shared/EntityDropdown/EntityDropdown';
 import { LineageTab } from '../shared/tabs/Lineage/LineageTab';
 import { capitalizeFirstLetterOnly } from '../../shared/textUtil';
 import { DashboardStatsSummarySubHeader } from './profile/DashboardStatsSummarySubHeader';
-import { chartMatchedFieldRenderer } from '../chart/ChartSnippet';
+import { chartMatchedFieldRenderer } from '../chart/chartMatchedFieldRenderer';
 import { EmbedTab } from '../shared/tabs/Embed/EmbedTab';
 import EmbeddedProfile from '../shared/embed/EmbeddedProfile';
 import DataProductSection from '../shared/containers/profile/sidebar/DataProduct/DataProductSection';
