--- datahub-web-react/src/app/entity/dataset/search/highlights.ts
@@ -1,7 +1,0 @@
-export const FIELDS_TO_HIGHLIGHT = new Map();
-FIELDS_TO_HIGHLIGHT.set('fieldPaths', 'column');
-FIELDS_TO_HIGHLIGHT.set('fieldDescriptions', 'column description');
-FIELDS_TO_HIGHLIGHT.set('fieldTags', 'column tag');
-FIELDS_TO_HIGHLIGHT.set('editedFieldDescriptions', 'column description');
-FIELDS_TO_HIGHLIGHT.set('editedFieldTags', 'column tag');
-FIELDS_TO_HIGHLIGHT.set('fieldLabels', 'label');
