+++ datahub-web-react/src/app/search/context/constants.ts
@@ -79,3 +79,12 @@ export const MATCHED_FIELD_MAPPING = {
     [EntityType.Dashboard]: CHART_DASHBOARD_FIELD_MAPPING,
     DEFAULT: DEFAULT_FIELD_MAPPING,
 } as const;
+
+export const FIELDS_TO_HIGHLIGHT = new Map();
+FIELDS_TO_HIGHLIGHT.set('fieldPaths', 'column');
+FIELDS_TO_HIGHLIGHT.set('fieldDescriptions', 'column description');
+FIELDS_TO_HIGHLIGHT.set('fieldTags', 'column tag');
+FIELDS_TO_HIGHLIGHT.set('editedFieldDescriptions', 'column description');
+FIELDS_TO_HIGHLIGHT.set('editedFieldTags', 'column tag');
+FIELDS_TO_HIGHLIGHT.set('fieldLabels', 'label');
+FIELDS_TO_HIGHLIGHT.set('editedFieldGlossaryTerms', 'column term');
